# TraceQL Performance: Drop Clone